refs-be-files.c: rename refs to refs-be-files
Rename refs.c to refs-be-files.c to indicate that this file now
holds the implementation for the files based refs backend.
A smaller portion of the code in this file is backend agnostic and will
be moved to a a new refs.c file that will hold all the common refs code
that is shared across all backends.

A second reason for first moving all the code to the new file and then move
the backend agnostic code back to refs.c instead of the other way around
is because the code that will eventually remain in this new refs-be-files.c
file is so entangled that it would then be very difficult to break the split
up into small independent patches/chunks.

Change-Id: I7a381773e24b1f54c8131f304c9b61b61a574e42
Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>